Update: RealCrossing: Modularization and cleanup.
