+++ src/main/java/org/opennars/applications/crossing/RealCrossing/MapEvidence.java
@@ -1,0 +1,46 @@
+/*
+ * The MIT License
+ *
+ * Copyright 2019 OpenNARS.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package org.opennars.applications.crossing.RealCrossing;
+
+import static org.opennars.applications.crossing.RealCrossing.TrafficMultiNar.locationNar;
+import org.opennars.entity.TruthValue;
+
+/**
+ *
+ * @author tc
+ */
+    public class MapEvidence {
+        public TruthValue car = new TruthValue(1.0f, 0.001f,locationNar.narParameters);
+        public TruthValue pedestrian = new TruthValue(1.0f, 0.001f,locationNar.narParameters);
+        public TruthValue bike = new TruthValue(1.0f, 0.001f,locationNar.narParameters);
+        public String choice() {
+            if(bike.getExpectation() > pedestrian.getExpectation() && bike.getExpectation() > car.getExpectation()) {
+                return "street"; //TODO bikelane
+            }
+            if(pedestrian.getExpectation() > bike.getExpectation() && pedestrian.getExpectation() > car.getExpectation()) {
+                return "sidewalk";
+            }
+            return "street";
+        }
+    }
